requirements.txt
langchain
openai
python-dotenv
pypdf
pdfminer
scikit-learn
sentence_transformers
google-generativeai
python-dateutil
langchain_openai
langchain-community